Declare global arch arrays with contents
This eliminates the need for archs_enable() and eliminates the racey
initialization.

This makes the architecture-specific init, option, and destroy functions
non-static so that they may be called from a different file.

The globals are now stored in cs_arch_globals.c, which is
generated by generate_cs_arch_globals.py. The
generate_cs_arch_globals.py script is not called by the build system
automatically, so developers must take care not to modify
cs_arch_globals.c manually.